+++ CHANGELOG.md
@@ -3,3 +3,4 @@
 01. Added dev.orne.config.Config
 01. Added dev.orne.config.AbstractConfig
 01. Added dev.orne.config.HierarchicalConfig
+01. Added dev.orne.config.AbstractStringConfig